+++ lsbi/model.py
@@ -2,235 +2,394 @@
 import numpy as np
 from numpy.linalg import inv, solve
 
-from lsbi.stats import (
-    mixture_multivariate_normal,
-    multimultivariate_normal,
-    multivariate_normal,
-)
-from lsbi.utils import logdet
+from lsbi.stats_1 import mixture_normal, multivariate_normal
+from lsbi.utils import logdet, matrix
+
+
+def _de_diagonalise(x, diagonal, *args):
+    if diagonal:
+        return np.atleast_1d(x)[..., None, :] * np.eye(*args)
+    else:
+        return x
 
 
 class LinearModel(object):
-    """A linear model.
+    """A multilinear model.
 
     D|theta ~ N( m + M theta, C )
     theta   ~ N( mu, Sigma )
 
     Defined by:
-        Parameters:       theta (n,)
-        Data:             D     (d,)
-        Prior mean:       mu    (n,)
-        Prior covariance: Sigma (n, n)
-        Data mean:        m     (d,)
-        Data covariance:  C     (d, d)
+        Parameters:       theta (..., n,)
+        Data:             D     (..., d,)
+        Model:            M     (..., d, n)
+        Prior mean:       mu    (..., n,)
+        Prior covariance: Sigma (..., n, n)
+        Data mean:        m     (..., d,)
+        Data covariance:  C     (..., d, d)
+
+    where the ellipses indicate arbitrary (broadcastable) additional copies.
 
     Parameters
     ----------
     M : array_like, optional
-        if matrix: model matrix
-        if vector: diagonal matrix with vector on diagonal
-        if scalar: scalar * rectangular identity matrix
+        if ndim>=2: model matrices
+        if ndim==1: model matrix with vector diagonal for all components
+        if ndim==0: scalar * rectangular identity matrix for all components
         Defaults to rectangular identity matrix
     m : array_like, optional
-        if vector: data mean
-        if scalar: scalar * unit vector
-        Defaults to zero vector
+        if ndim>=1: data means
+        if ndim==0: scalar * unit vector for all components
+        Defaults to 0 for all components
     C : array_like, optional
-        if matrix: data covariance
-        if vector: diagonal matrix with vector on diagonal
-        if scalar: scalar * identity matrix
-        Defaults to identity matrix
+        if ndim>=2: data covariances
+        if ndim==1: data covariance with vector diagonal for all components
+        if ndim==0: scalar * identity matrix for all components
+        Defaults to rectangular identity matrix
     mu : array_like, optional
-        if vector: prior mean
-        if scalar: scalar * unit vector
-        Defaults to zero vector
+        if ndim>=1: prior means
+        if ndim==0: scalar * unit vector for all components
+        Defaults to 0 for all components
+        Prior mean, defaults to zero vector
     Sigma : array_like, optional
-        if matrix: prior covariance
-        if vector: diagonal matrix with vector on diagonal
-        if scalar: scalar * identity matrix
-        Defaults to identity matrix
+        if ndim>=2: prior covariances
+        if ndim==1: prior covariance with vector diagonal for all components
+        if ndim==0: scalar * identity matrix for all components
+        Defaults to k copies of identity matrices
     n : int, optional
-        Number of parameters
-        Defaults to automatically inferred value
+        Number of parameters, defaults to automatically inferred value
     d : int, optional
-        Number of data dimensions
-        Defaults to automatically inferred value
+        Number of data dimensions, defaults to automatically inferred value
+    shape : (), optional
+        Number of mixture components, defaults to automatically inferred value
     """
 
-    def __init__(self, *args, **kwargs):
-        # Rationalise input arguments
-        M = self._atleast_2d(kwargs.pop("M", None))
-        m = self._atleast_1d(kwargs.pop("m", None))
-        C = self._atleast_2d(kwargs.pop("C", None))
-        mu = self._atleast_1d(kwargs.pop("mu", None))
-        Sigma = self._atleast_2d(kwargs.pop("Sigma", None))
-        n = kwargs.pop("n", 0)
-        d = kwargs.pop("d", 0)
+    def __init__(
+        self,
+        M=1,
+        m=0,
+        C=1,
+        mu=0,
+        Sigma=1,
+        shape=(),
+        n=1,
+        d=1,
+        diagonal_M=False,
+        diagonal_C=False,
+        diagonal_Sigma=False,
+    ):
+        self.M = M
+        self.diagonal_M = diagonal_M
+        if len(np.shape(self.M)) < 2:
+            self.diagonal_M = True
+        self.m = m
+        self.C = C
+        self.diagonal_C = diagonal_C
+        if len(np.shape(self.C)) < 2:
+            self.diagonal_C = True
+        self.mu = mu
+        self.Sigma = Sigma
+        self.diagonal_Sigma = diagonal_Sigma
+        if len(np.shape(self.Sigma)) < 2:
+            self.diagonal_Sigma = True
+        self._shape = shape
+        self._n = n
+        self._d = d
 
-        # Determine dimensions
-        n = max([n, M.shape[1], mu.shape[0], Sigma.shape[0], Sigma.shape[1]])
-        d = max([d, M.shape[0], m.shape[0], C.shape[0], C.shape[1]])
-        if not n:
-            raise ValueError("Unable to determine number of parameters n")
-        if not d:
-            raise ValueError("Unable to determine data dimensions d")
-
-        # Set defaults if no argument was passed
-        M = M if M.size else np.eye(d, n)
-        m = m if m.size else np.zeros(d)
-        C = C if C.size else np.eye(d)
-        mu = mu if mu.size else np.zeros(n)
-        Sigma = Sigma if Sigma.size else np.eye(n)
-
-        # Broadcast to correct shape
-        self.M = self._broadcast_to(M, (d, n))
-        self.m = np.broadcast_to(m, (d,))
-        self.C = self._broadcast_to(C, (d, d))
-        self.mu = np.broadcast_to(mu, (n,))
-        self.Sigma = self._broadcast_to(Sigma, (n, n))
-
-    @classmethod
-    def from_joint(cls, mean, cov, n):
-        """Construct model from joint distribution."""
-        mean = np.atleast_1d(mean)
-        cov = np.atleast_2d(cov)
-        mu = mean[-n:]
-        Sigma = cov[-n:, -n:]
-        M = solve(Sigma, cov[-n:, :-n]).T
-        m = mean[:-n] - M @ mu
-        C = cov[:-n, :-n] - M @ Sigma @ M.T
-
-        return cls(M=M, m=m, C=C, mu=mu, Sigma=Sigma)
+    @property
+    def shape(self):
+        """Shape of the distribution."""
+        return np.broadcast_shapes(
+            np.shape(self.M)[: -2 + self.diagonal_M],
+            np.shape(self.m)[:-1],
+            np.shape(self.C)[: -2 + self.diagonal_C],
+            np.shape(self.mu)[:-1],
+            np.shape(self.Sigma)[: -2 + self.diagonal_Sigma],
+            self._shape,
+        )
 
     @property
     def n(self):
-        """Dimensionality of parameter space len(theta)."""
-        return self.M.shape[1]
+        """Dimension of the distribution."""
+        return np.max(
+            [
+                *np.shape(self.M)[len(np.shape(self.M)) - 1 + self.diagonal_M :],
+                *np.shape(self.Sigma)[-2 + self.diagonal_Sigma :],
+                *np.shape(self.mu)[-1:],
+                self._n,
+            ]
+        )
 
     @property
     def d(self):
         """Dimensionality of data space len(D)."""
-        return self.M.shape[0]
+        return np.max(
+            [
+                *np.shape(self.M)[-2 + self.diagonal_M : -1],
+                *np.shape(self.C)[-2 + self.diagonal_C :],
+                *np.shape(self.m)[-1:],
+                self._d,
+            ]
+        )
 
     def likelihood(self, theta):
         """P(D|theta) as a scipy distribution object.
 
-        D ~ N( m + M theta, C )
+        D|theta ~ N( m + M theta, C )
+        theta   ~ N( mu, Sigma )
 
         Parameters
         ----------
-        theta : array_like, shape (n,)
+        theta : array_like, shape (k, n)
         """
-        theta = np.atleast_1d(theta)
-        return multivariate_normal(self.m + self.M @ theta, self.C)
+        mu = self.m + np.einsum("...ja,...a->...j", self._M, theta)
+        return multivariate_normal(mu, self.C, self.shape, self.d, self.diagonal_C)
 
     def prior(self):
         """P(theta) as a scipy distribution object.
 
         theta ~ N( mu, Sigma )
         """
-        return multivariate_normal(self.mu, self.Sigma)
+        return multivariate_normal(
+            self.mu, self.Sigma, self.shape, self.n, self.diagonal_Sigma
+        )
 
     def posterior(self, D):
         """P(theta|D) as a scipy distribution object.
 
-        theta ~ N( mu + Sigma M'C^{-1}(D-m), Sigma - Sigma M' C^{-1} M Sigma )
+        theta|D ~ N( mu + S M'C^{-1}(D - m - M mu), S )
+        S = (Sigma^{-1} + M'C^{-1}M)^{-1}
 
         Parameters
         ----------
         D : array_like, shape (d,)
         """
-        D = np.atleast_1d(D)
-        Sigma = inv(inv(self.Sigma) + self.M.T @ inv(self.C) @ self.M)
-        D0 = self.m + self.M @ self.mu
-        mu = self.mu + Sigma @ self.M.T @ inv(self.C) @ (D - D0)
-        return multivariate_normal(mu, Sigma)
+        values = (
+            D
+            - self.m
+            - np.einsum("...ja,...a->...j", self._M, self.mu * np.ones(self.n))
+        )
+
+        diagonal_Sigma = self.diagonal_C and self.diagonal_Sigma and self.diagonal_M
+
+        if diagonal_Sigma:
+            dim = min(self.n, self.d)
+            shape = np.broadcast_shapes(self.shape, values.shape[:-1])
+            C = np.atleast_1d(self.C)[..., :dim]
+            M = np.atleast_1d(self.M)[..., :dim]
+            Sigma = np.broadcast_to(self.Sigma, shape + (self.n,)).copy()
+            Sigma[..., :dim] = 1 / (1 / Sigma[..., :dim] + M**2 / C)
+
+            mu = np.broadcast_to(self.mu, shape + (self.n,)).copy()
+            mu[..., :dim] = mu[..., :dim] + Sigma[..., :dim] * M / C * values[..., :dim]
+        else:
+            if self.diagonal_C:
+                invC = np.eye(self.d) / np.atleast_1d(self.C)[..., None, :]
+            else:
+                invC = inv(self.C)
+
+            if self.diagonal_Sigma:
+                invSigma = np.eye(self.n) / np.atleast_1d(self.Sigma)[..., None, :]
+            else:
+                invSigma = inv(self.Sigma)
+
+            Sigma = inv(
+                invSigma + np.einsum("...aj,...ab,...bk->...jk", self._M, invC, self._M)
+            )
+            mu = self.mu + np.einsum(
+                "...ja,...ba,...bc,...c->...j", Sigma, self._M, invC, values
+            )
+
+        return multivariate_normal(mu, Sigma, self.shape, self.n, diagonal_Sigma)
 
     def evidence(self):
         """P(D) as a scipy distribution object.
 
         D ~ N( m + M mu, C + M Sigma M' )
         """
-        return multivariate_normal(
-            self.m + self.M @ self.mu, self.C + self.M @ self.Sigma @ self.M.T
-        )
+        mu = self.m + np.einsum("...ja,...a->...j", self._M, self.mu * np.ones(self.n))
+        diagonal_Sigma = self.diagonal_C and self.diagonal_Sigma and self.diagonal_M
+
+        if diagonal_Sigma:
+            dim = min(self.n, self.d)
+            M = np.atleast_1d(self.M)[..., :dim]
+            S = np.atleast_1d(self.Sigma)[..., :dim]
+            Sigma = np.broadcast_to(self.C, self.shape + (self.d,)).copy()
+            Sigma[..., :dim] += S * M**2
+        else:
+            Sigma = self._C + np.einsum(
+                "...ja,...ab,...kb->...jk", self._M, self._Sigma, self._M
+            )
+
+        return multivariate_normal(mu, Sigma, self.shape, self.d, diagonal_Sigma)
 
     def joint(self):
         """P(D, theta) as a scipy distribution object.
 
-        [  D  ] ~ N( [m + M mu]   [C + M Sigma M'  M Sigma] )
-        [theta]    ( [   mu   ] , [   Sigma M'      Sigma ] )
+        [  D  ] | A ~ N( [m + M mu]   [C + M Sigma M'  M Sigma] )
+        [theta] |      ( [   mu   ] , [   Sigma M'      Sigma ] )
         """
         evidence = self.evidence()
         prior = self.prior()
-        mu = np.concatenate([evidence.mean, prior.mean])
-        Sigma = np.block(
-            [[evidence.cov, self.M @ self.Sigma], [self.Sigma @ self.M.T, prior.cov]]
-        )
-        return multivariate_normal(mu, Sigma)
+        a = np.broadcast_to(evidence.mean, self.shape + (self.d,))
+        b = np.broadcast_to(prior.mean, self.shape + (self.n,))
+        mu = np.block([a, b])
+        A = _de_diagonalise(evidence.cov, evidence.diagonal_cov, self.d)
+        A = np.broadcast_to(A, self.shape + (self.d, self.d))
+        D = _de_diagonalise(prior.cov, prior.diagonal_cov, self.n)
+        D = np.broadcast_to(D, self.shape + (self.n, self.n))
+        B = np.einsum("...ja,...al->...jl", self._M, self._Sigma)
+        B = np.broadcast_to(B, self.shape + (self.d, self.n))
+        C = np.moveaxis(B, -1, -2)
+        Sigma = np.block([[A, B], [C, D]])
+        return multivariate_normal(mu, Sigma, self.shape, self.n + self.d)
 
-    def DKL(self, D):
-        """D_KL(P(theta|D)||P(theta)) the Kullback-Leibler divergence.
+    @property
+    def _M(self):
+        return _de_diagonalise(self.M, self.diagonal_M, self.d, self.n)
+
+    @property
+    def _C(self):
+        return _de_diagonalise(self.C, self.diagonal_C, self.d)
+
+    @property
+    def _Sigma(self):
+        return _de_diagonalise(self.Sigma, self.diagonal_Sigma, self.n)
+
+
+class MixtureModel(LinearModel):
+    """A linear mixture model.
+
+    D|theta, A ~ N( m + M theta, C )
+    theta|A    ~ N( mu, Sigma )
+    A          ~ categorical( exp(logA) )
+
+    Defined by:
+        Parameters:          theta (..., n,)
+        Data:                D     (..., d,)
+        Prior means:         mu    (..., k, n)
+        Prior covariances:   Sigma (..., k, n, n)
+        Data means:          m     (..., k, d)
+        Data covariances:    C     (..., k, d, d)
+        log mixture weights: logA  (..., k,)
+
+    Parameters
+    ----------
+    M : array_like, optional
+        if ndim>=2: model matrices
+        if ndim==1: model matrix with vector diagonal for all components
+        if scalar: scalar * rectangular identity matrix for all components
+        Defaults to k copies of rectangular identity matrices
+    m : array_like, optional
+        if ndim>=1: data means
+        if scalar: scalar * unit vector for all components
+        Defaults to 0 for all components
+    C : array_like, optional
+        if ndim>=2: data covariances
+        if ndim==1: data covariance with vector diagonal for all components
+        if scalar: scalar * identity matrix for all components
+        Defaults to k copies of identity matrices
+    mu : array_like, optional
+        if ndim>=1: prior means
+        if scalar: scalar * unit vector for all components
+        Defaults to 0 for all components
+        Prior mean, defaults to zero vector
+    Sigma : array_like, optional
+        if ndim>=2: prior covariances
+        if ndim==1: prior covariance with vector diagonal for all components
+        if scalar: scalar * identity matrix for all components
+        Defaults to k copies of identity matrices
+    logA : array_like, optional
+        if ndim>=1: log mixture weights
+        if scalar: scalar * unit vector
+        Defaults to uniform weights
+    n : int, optional
+        Number of parameters, defaults to automatically inferred value
+    d : int, optional
+        Number of data dimensions, defaults to automatically inferred value
+    """
+
+    def __init__(self, logA=1, *args):
+        super().__init__(*args)
+        self.logA = logA
+
+    @property
+    def shape(self):
+        """Shape of the distribution."""
+        return np.broadcast_shapes(np.shape(self.logA), super().shape)
+
+    @property
+    def k(self):
+        """Number of mixture components of the distribution."""
+        return self.shape[-1]
+
+    def likelihood(self, theta):
+        """P(D|theta) as a scipy distribution object.
+
+        D|theta,A ~ N( m + M theta, C )
+        theta|A   ~ N( mu, Sigma )
+        A         ~ categorical(exp(logA))
+
+        Parameters
+        ----------
+        theta : array_like, shape (n,)
+        """
+        dist = super().likelihood(np.expand_dims(theta, -2))
+        dist.__class__ = mixture_normal
+        dist.logA = self.prior().logpdf(theta, broadcast=True, joint=True)
+        return dist
+
+    def prior(self):
+        """P(theta) as a scipy distribution object.
+
+        theta|A ~ N( mu, Sigma )
+        A       ~ categorical(exp(logA))
+        """
+        dist = super().prior()
+        dist.__class__ = mixture_normal
+        dist.logA = self.logA
+        return dist
+
+    def posterior(self, D):
+        """P(theta|D) as a scipy distribution object.
+
+        theta|D, A ~ N( mu + S M'C^{-1}(D - m - M mu), S )
+        D|A        ~ N( m + M mu, C + M Sigma M' )
+        A          ~ categorical(exp(logA))
+        S = (Sigma^{-1} + M'C^{-1}M)^{-1}
 
         Parameters
         ----------
         D : array_like, shape (d,)
         """
-        cov_p = self.posterior(D).cov
-        cov_q = self.prior().cov
-        mu_p = self.posterior(D).mean
-        mu_q = self.prior().mean
-        return (
-            -logdet(cov_p)
-            + logdet(cov_q)
-            + np.trace(inv(cov_q) @ cov_p - 1)
-            + (mu_q - mu_p) @ inv(cov_q) @ (mu_q - mu_p)
-        ) / 2
+        dist = super().posterior(np.expand_dims(D, -2))
+        dist.__class__ = mixture_normal
+        dist.logA = self.evidence().logpdf(D, broadcast=True, joint=True)
+        return dist
 
-    def reduce(self, D):
-        """Reduce the model to a Gaussian in the parameters.
+    def evidence(self):
+        """P(D) as a scipy distribution object.
 
-        Parameters
-        ----------
-        D : array_like, shape (d,)
-
-        Returns
-        -------
-        ReducedLinearModel
+        D|A ~ N( m + M mu, C + M Sigma M' )
+        A   ~ categorical(exp(logA))
         """
-        Sigma_L = inv(self.M.T @ inv(self.C) @ self.M)
-        mu_L = Sigma_L @ self.M.T @ inv(self.C) @ (D - self.m)
-        logLmax = (
-            -logdet(2 * np.pi * self.C) / 2
-            - (D - self.m)
-            @ inv(self.C)
-            @ (self.C - self.M @ Sigma_L @ self.M.T)
-            @ inv(self.C)
-            @ (D - self.m)
-            / 2
-        )
-        return ReducedLinearModel(
-            mu_L=mu_L,
-            Sigma_L=Sigma_L,
-            logLmax=logLmax,
-            mu_pi=self.prior().mean,
-            Sigma_pi=self.prior().cov,
-        )
+        dist = super().evidence()
+        dist.__class__ = mixture_normal
+        dist.logA = self.logA
+        return dist
 
-    def _atleast_2d(self, x):
-        if x is None:
-            return np.zeros(shape=(0, 0))
-        return np.atleast_2d(x)
+    def joint(self):
+        """P(D, theta) as a scipy distribution object.
 
-    def _atleast_1d(self, x):
-        if x is None:
-            return np.zeros(shape=(0,))
-        return np.atleast_1d(x)
+        [  D  ] | A ~ N( [m + M mu]   [C + M Sigma M'  M Sigma] )
+        [theta] |      ( [   mu   ] , [   Sigma M'      Sigma ] )
 
-    def _broadcast_to(self, x, shape):
-        if x.shape == shape:
-            return x
-        return x * np.eye(*shape)
+        A           ~ categorical(exp(logA))
+        """
+        dist = super().joint()
+        dist.__class__ = mixture_normal
+        dist.logA = self.logA
+        return dist
 
 
 class ReducedLinearModel(object):
@@ -383,449 +542,3 @@ class ReducedLinearModelUniformPrior(object):
     def DKL(self):
         """D_KL(P(theta|D)||P(theta)) the Kullback-Leibler divergence."""
         return self.logV - logdet(2 * np.pi * np.e * self.Sigma_P) / 2
-
-
-class LinearMixtureModel(object):
-    """A linear mixture model.
-
-    D|theta, A ~ N( m + M theta, C )
-    theta|A    ~ N( mu, Sigma )
-    A          ~ categorical( exp(logA) )
-
-    Defined by:
-        Parameters:          theta (n,)
-        Data:                D     (d,)
-        Prior means:         mu    (k, n)
-        Prior covariances:   Sigma (k, n, n)
-        Data means:          m     (k, d)
-        Data covariances:    C     (k, d, d)
-        log mixture weights: logA  (k,)
-
-    Parameters
-    ----------
-    M : array_like, optional
-        if ndim==3: model matrices
-        if ndim==2: model matrix with same matrix for all components
-        if ndim==1: model matrix with vector diagonal for all components
-        if scalar: scalar * rectangular identity matrix for all components
-        Defaults to k copies of rectangular identity matrices
-    m : array_like, optional
-        if ndim==2: data means
-        if ndim==1: data mean with same vector for all components
-        if scalar: scalar * unit vector for all components
-        Defaults to 0 for all components
-    C : array_like, optional
-        if ndim==3: data covariances
-        if ndim==2: data covariance with same matrix for all components
-        if ndim==1: data covariance with vector diagonal for all components
-        if scalar: scalar * identity matrix for all components
-        Defaults to k copies of identity matrices
-    mu : array_like, optional
-        if ndim==2: prior means
-        if ndim==1: prior mean with same vector for all components
-        if scalar: scalar * unit vector for all components
-        Defaults to 0 for all components
-        Prior mean, defaults to zero vector
-    Sigma : array_like, optional
-        if ndim==3: prior covariances
-        if ndim==2: prior covariance with same matrix for all components
-        if ndim==1: prior covariance with vector diagonal for all components
-        if scalar: scalar * identity matrix for all components
-        Defaults to k copies of identity matrices
-    logA : array_like, optional
-        if ndim==1: log mixture weights
-        if scalar: scalar * unit vector
-        Defaults to uniform weights
-    n : int, optional
-        Number of parameters, defaults to automatically inferred value
-    d : int, optional
-        Number of data dimensions, defaults to automatically inferred value
-    k : int, optional
-        Number of mixture components, defaults to automatically inferred value
-    """
-
-    def __init__(self, *args, **kwargs):
-        # Rationalise input arguments
-        M = self._atleast_3d(kwargs.pop("M", None))
-        m = self._atleast_2d(kwargs.pop("m", None))
-        C = self._atleast_3d(kwargs.pop("C", None))
-        mu = self._atleast_2d(kwargs.pop("mu", None))
-        Sigma = self._atleast_3d(kwargs.pop("Sigma", None))
-        logA = self._atleast_1d(kwargs.pop("logA", None))
-        n = kwargs.pop("n", 0)
-        d = kwargs.pop("d", 0)
-        k = kwargs.pop("k", 0)
-
-        # Determine dimensions
-        n = max([n, M.shape[2], mu.shape[1], Sigma.shape[1], Sigma.shape[2]])
-        d = max([d, M.shape[1], m.shape[1], C.shape[1], C.shape[2]])
-        k = max(
-            [
-                k,
-                M.shape[0],
-                m.shape[0],
-                C.shape[0],
-                mu.shape[0],
-                Sigma.shape[0],
-                logA.shape[0],
-            ]
-        )
-        if not n:
-            raise ValueError("Unable to determine number of parameters n")
-        if not d:
-            raise ValueError("Unable to determine data dimensions d")
-
-        # Set defaults if no argument was passed
-        M = M if M.size else np.eye(d, n)
-        m = m if m.size else np.zeros(d)
-        C = C if C.size else np.eye(d)
-        mu = mu if mu.size else np.zeros(n)
-        Sigma = Sigma if Sigma.size else np.eye(n)
-        logA = logA if logA.size else -np.log(k)
-
-        # Broadcast to correct shape
-        self.M = self._broadcast_to(M, (k, d, n))
-        self.m = np.broadcast_to(m, (k, d))
-        self.C = self._broadcast_to(C, (k, d, d))
-        self.mu = np.broadcast_to(mu, (k, n))
-        self.Sigma = self._broadcast_to(Sigma, (k, n, n))
-        self.logA = np.broadcast_to(logA, (k,))
-
-    @classmethod
-    def from_joint(cls, means, covs, logA, n):
-        """Construct model from joint distribution."""
-        mu = means[:, -n:]
-        Sigma = covs[:, -n:, -n:]
-        M = solve(Sigma, covs[:, -n:, :-n]).transpose(0, 2, 1)
-        m = means[:, :-n] - np.einsum("ija,ia->ij", M, mu)
-        C = covs[:, :-n, :-n] - np.einsum("ija,iab,ikb->ijk", M, Sigma, M)
-        return cls(M=M, m=m, C=C, mu=mu, Sigma=Sigma, logA=logA)
-
-    @property
-    def n(self):
-        """Dimensionality of parameter space len(theta)."""
-        return self.M.shape[2]
-
-    @property
-    def d(self):
-        """Dimensionality of data space len(D)."""
-        return self.M.shape[1]
-
-    @property
-    def k(self):
-        """Number of mixture components len(logA)."""
-        return self.M.shape[0]
-
-    def likelihood(self, theta):
-        """P(D|theta) as a scipy distribution object.
-
-        D|theta,A ~ N( m + M theta, C )
-        theta|A   ~ N( mu, Sigma )
-        A         ~ categorical(exp(logA))
-
-        Parameters
-        ----------
-        theta : array_like, shape (n,)
-        """
-        theta = np.atleast_1d(theta)
-        mu = self.m + np.einsum("ija,a->ij", self.M, theta)
-        prior = self.prior()
-        logA = prior.logpdf(theta, reduce=False) + self.logA - prior.logpdf(theta)
-        return mixture_multivariate_normal(mu, self.C, logA)
-
-    def prior(self):
-        """P(theta) as a scipy distribution object.
-
-        theta|A ~ N( mu, Sigma )
-        A       ~ categorical(exp(logA))
-        """
-        return mixture_multivariate_normal(self.mu, self.Sigma, self.logA)
-
-    def posterior(self, D):
-        """P(theta|D) as a scipy distribution object.
-
-        theta|D, A ~ N( mu + S M'C^{-1}(D - m - M mu), S )
-        D|A        ~ N( m + M mu, C + M Sigma M' )
-        A          ~ categorical(exp(logA))
-        S = (Sigma^{-1} + M'C^{-1}M)^{-1}
-
-        Parameters
-        ----------
-        D : array_like, shape (d,)
-        """
-        D = np.atleast_1d(D)
-        Sigma = inv(
-            inv(self.Sigma) + np.einsum("iaj,iab,ibk->ijk", self.M, inv(self.C), self.M)
-        )
-        D0 = self.m + np.einsum("ija,ia->ij", self.M, self.mu)
-        mu = self.mu + np.einsum(
-            "ija,iba,ibc,ic->ij", Sigma, self.M, inv(self.C), D - D0
-        )
-        evidence = self.evidence()
-        logA = evidence.logpdf(D, reduce=False) + self.logA - evidence.logpdf(D)
-        return mixture_multivariate_normal(mu, Sigma, logA)
-
-    def evidence(self):
-        """P(D) as a scipy distribution object.
-
-        D|A ~ N( m + M mu, C + M Sigma M' )
-        A   ~ categorical(exp(logA))
-        """
-        mu = self.m + np.einsum("ija,ia->ij", self.M, self.mu)
-        Sigma = self.C + np.einsum("ija,iab,ikb->ijk", self.M, self.Sigma, self.M)
-        return mixture_multivariate_normal(mu, Sigma, self.logA)
-
-    def joint(self):
-        """P(D, theta) as a scipy distribution object.
-
-        [  D  ] | A ~ N( [m + M mu]   [C + M Sigma M'  M Sigma] )
-        [theta] |      ( [   mu   ] , [   Sigma M'      Sigma ] )
-
-        A           ~ categorical(exp(logA))
-        """
-        evidence = self.evidence()
-        prior = self.prior()
-        mu = np.block([evidence.means, prior.means])
-        corr = np.einsum("ija,ial->ijl", self.M, self.Sigma)
-        Sigma = np.block([[evidence.covs, corr], [corr.transpose(0, 2, 1), prior.covs]])
-        return mixture_multivariate_normal(mu, Sigma, self.logA)
-
-    def _atleast_3d(self, x):
-        if x is None:
-            return np.zeros(shape=(0, 0, 0))
-        x = np.array(x)
-        if x.ndim == 3:
-            return x
-        return np.atleast_2d(x)[None, ...]
-
-    def _atleast_2d(self, x):
-        if x is None:
-            return np.zeros(shape=(0, 0))
-        x = np.array(x)
-        if x.ndim == 2:
-            return x
-        return np.atleast_1d(x)[None, ...]
-
-    def _atleast_1d(self, x):
-        if x is None:
-            return np.zeros(shape=(0,))
-        return np.atleast_1d(x)
-
-    def _broadcast_to(self, x, shape):
-        if x.shape == shape:
-            return x
-        if x.shape[1:] == shape[1:]:
-            return np.broadcast_to(x, shape)
-        return x * np.ones(shape) * np.eye(shape[1], shape[2])[None, ...]
-
-
-class MultiLinearModel(object):
-    """A multilinear model.
-
-    D|theta ~ N( m + M theta, C )
-    theta   ~ N( mu, Sigma )
-
-    Defined by:
-        Parameters:       theta (k, n,)
-        Data:             D     (k, d,)
-        Prior mean:       mu    (k, n,)
-        Prior covariance: Sigma (k, n, n)
-        Data mean:        m     (k, d,)
-        Data covariance:  C     (k, d, d)
-
-    i.e. the same as a LinearModel, but with k copies of each parameter.
-    Fully vectorised so k may in principle be very large.
-
-    Parameters
-    ----------
-    M : array_like, optional
-        if ndim==3: model matrices
-        if ndim==2: model matrix with same matrix for all components
-        if ndim==1: model matrix with vector diagonal for all components
-        if scalar: scalar * rectangular identity matrix for all components
-        Defaults to k copies of rectangular identity matrices
-    m : array_like, optional
-        if ndim==2: data means
-        if ndim==1: data mean with same vector for all components
-        if scalar: scalar * unit vector for all components
-        Defaults to 0 for all components
-    C : array_like, optional
-        if ndim==3: data covariances
-        if ndim==2: data covariance with same matrix for all components
-        if ndim==1: data covariance with vector diagonal for all components
-        if scalar: scalar * identity matrix for all components
-        Defaults to k copies of identity matrices
-    mu : array_like, optional
-        if ndim==2: prior means
-        if ndim==1: prior mean with same vector for all components
-        if scalar: scalar * unit vector for all components
-        Defaults to 0 for all components
-        Prior mean, defaults to zero vector
-    Sigma : array_like, optional
-        if ndim==3: prior covariances
-        if ndim==2: prior covariance with same matrix for all components
-        if ndim==1: prior covariance with vector diagonal for all components
-        if scalar: scalar * identity matrix for all components
-        Defaults to k copies of identity matrices
-    n : int, optional
-        Number of parameters, defaults to automatically inferred value
-    d : int, optional
-        Number of data dimensions, defaults to automatically inferred value
-    k : int, optional
-        Number of mixture components, defaults to automatically inferred value
-    """
-
-    def __init__(self, *args, **kwargs):
-        # Rationalise input arguments
-        M = self._atleast_3d(kwargs.pop("M", None))
-        m = self._atleast_2d(kwargs.pop("m", None))
-        C = self._atleast_3d(kwargs.pop("C", None))
-        mu = self._atleast_2d(kwargs.pop("mu", None))
-        Sigma = self._atleast_3d(kwargs.pop("Sigma", None))
-        n = kwargs.pop("n", 0)
-        d = kwargs.pop("d", 0)
-        k = kwargs.pop("k", 0)
-
-        # Determine dimensions
-        n = max([n, M.shape[2], mu.shape[1], Sigma.shape[1], Sigma.shape[2]])
-        d = max([d, M.shape[1], m.shape[1], C.shape[1], C.shape[2]])
-        k = max(
-            [
-                k,
-                M.shape[0],
-                m.shape[0],
-                C.shape[0],
-                mu.shape[0],
-                Sigma.shape[0],
-            ]
-        )
-        if not n:
-            raise ValueError("Unable to determine number of parameters n")
-        if not d:
-            raise ValueError("Unable to determine data dimensions d")
-
-        # Set defaults if no argument was passed
-        M = M if M.size else np.eye(d, n)
-        m = m if m.size else np.zeros(d)
-        C = C if C.size else np.eye(d)
-        mu = mu if mu.size else np.zeros(n)
-        Sigma = Sigma if Sigma.size else np.eye(n)
-
-        # Broadcast to correct shape
-        self.M = self._broadcast_to(M, (k, d, n))
-        self.m = np.broadcast_to(m, (k, d))
-        self.C = self._broadcast_to(C, (k, d, d))
-        self.mu = np.broadcast_to(mu, (k, n))
-        self.Sigma = self._broadcast_to(Sigma, (k, n, n))
-
-    @classmethod
-    def from_joint(cls, means, covs, n):
-        """Construct model from joint distribution."""
-        mu = means[:, -n:]
-        Sigma = covs[:, -n:, -n:]
-        M = solve(Sigma, covs[:, -n:, :-n]).transpose(0, 2, 1)
-        m = means[:, :-n] - np.einsum("ija,ia->ij", M, mu)
-        C = covs[:, :-n, :-n] - np.einsum("ija,iab,ikb->ijk", M, Sigma, M)
-        return cls(M=M, m=m, C=C, mu=mu, Sigma=Sigma)
-
-    @property
-    def n(self):
-        """Dimensionality of parameter space len(theta)."""
-        return self.M.shape[2]
-
-    @property
-    def d(self):
-        """Dimensionality of data space len(D)."""
-        return self.M.shape[1]
-
-    @property
-    def k(self):
-        """Number of copies."""
-        return self.M.shape[0]
-
-    def likelihood(self, theta):
-        """P(D|theta) as a scipy distribution object.
-
-        D|theta ~ N( m + M theta, C )
-        theta   ~ N( mu, Sigma )
-
-        Parameters
-        ----------
-        theta : array_like, shape (k, n)
-        """
-        theta = np.array(theta).reshape(self.k, self.n)
-        mu = self.m + np.einsum("ija,ia->ij", self.M, theta)
-        return multimultivariate_normal(mu, self.C)
-
-    def prior(self):
-        """P(theta) as a scipy distribution object.
-
-        theta ~ N( mu, Sigma )
-        """
-        return multimultivariate_normal(self.mu, self.Sigma)
-
-    def posterior(self, D):
-        """P(theta|D) as a scipy distribution object.
-
-        theta|D ~ N( mu + S M'C^{-1}(D - m - M mu), S )
-        D       ~ N( m + M mu, C + M Sigma M' )
-        S = (Sigma^{-1} + M'C^{-1}M)^{-1}
-
-        Parameters
-        ----------
-        D : array_like, shape (d,)
-        """
-        D = D.reshape(self.k, self.d)
-        Sigma = inv(
-            inv(self.Sigma) + np.einsum("iaj,iab,ibk->ijk", self.M, inv(self.C), self.M)
-        )
-        D0 = self.m + np.einsum("ija,ia->ij", self.M, self.mu)
-        mu = self.mu + np.einsum(
-            "ija,iba,ibc,ic->ij", Sigma, self.M, inv(self.C), D - D0
-        )
-        return multimultivariate_normal(mu, Sigma)
-
-    def evidence(self):
-        """P(D) as a scipy distribution object.
-
-        D|A ~ N( m + M mu, C + M Sigma M' )
-        """
-        mu = self.m + np.einsum("ija,ia->ij", self.M, self.mu)
-        Sigma = self.C + np.einsum("ija,iab,ikb->ijk", self.M, self.Sigma, self.M)
-        return multimultivariate_normal(mu, Sigma)
-
-    def joint(self):
-        """P(D, theta) as a scipy distribution object.
-
-        [  D  ] | A ~ N( [m + M mu]   [C + M Sigma M'  M Sigma] )
-        [theta] |      ( [   mu   ] , [   Sigma M'      Sigma ] )
-        """
-        evidence = self.evidence()
-        prior = self.prior()
-        mu = np.block([evidence.means, prior.means])
-        corr = np.einsum("ija,ial->ijl", self.M, self.Sigma)
-        Sigma = np.block([[evidence.covs, corr], [corr.transpose(0, 2, 1), prior.covs]])
-        return multimultivariate_normal(mu, Sigma)
-
-    def _atleast_3d(self, x):
-        if x is None:
-            return np.zeros(shape=(0, 0, 0))
-        x = np.array(x)
-        if x.ndim == 3:
-            return x
-        return np.atleast_2d(x)[None, ...]
-
-    def _atleast_2d(self, x):
-        if x is None:
-            return np.zeros(shape=(0, 0))
-        x = np.array(x)
-        if x.ndim == 2:
-            return x
-        return np.atleast_1d(x)[None, ...]
-
-    def _broadcast_to(self, x, shape):
-        if x.shape == shape:
-            return x
-        if x.shape[1:] == shape[1:]:
-            return np.broadcast_to(x, shape)
-        return x * np.ones(shape) * np.eye(shape[1], shape[2])[None, ...]
